feat: skeleton table for streaming and loading
