+++ apps/web/app/environments/[environmentId]/people/Loading.tsx
@@ -1,0 +1,5 @@
+import LoadingTable from "@/components/shared/LoadingTable";
+
+export default function Loading() {
+    return <LoadingTable />
+}
