+++ apps/web/app/environments/[environmentId]/people/page.tsx
@@ -2,11 +2,11 @@ export const revalidate = 0;
 
 import { Suspense } from "react";
 import PeopleList from "@/app/environments/[environmentId]/people/PeopleList";
-import LoadingSpinnerInTable from "@/components/shared/LoadingSpinnerInTable";
+import Loading from "@/app/environments/[environmentId]/people/Loading";
 
 export default function PeoplePage({ params }) {
   return (
-    <Suspense fallback={<LoadingSpinnerInTable />}>
+    <Suspense fallback={<Loading />}>
       {/* @ts-expect-error Server Component */}
       <PeopleList environmentId={params.environmentId} />
     </Suspense>
